Add PKGBUILDS for a few Haskell programs
There has been a regression in the packaging of Haskell programs on
Arch recently, where packages such as pandoc now have hundreds of
dependencies, with all Hackage packages packaged individually. This is
bad for users, who don't want or need 700 MiB of Hackage stuff, and it
is bad for Haskell developers too, as the language package manager
(Stack or Cabal) is more suitable for tracking these dependencies
anyway. While it is true that dynamic linking allows for reuse across
dependencies, in this case the cost is far greater (hundreds of MiBs
and frequent useless new package releases) than the benefits (reusing
a few KiBs across the three Haskell packages I use). Furthermore, when
building with --split-objs there is even less overlap.

Stack pretty much solved this issue of building Haskell programs, and
all of the Haskell programs that I care about happen to support building
with Stack. So let's make a few PKGBUILDs that build these programs from
source using Stack, and package static binaries.